WEB-4460 abstract Edit Patient Dialog
